Make markChannelRead thread-safe
All channels' threads will now be able to mark the message without losing
potential changes.

In Example, If two threads go to update the read at the same time then its
possible that the second thread would not keep the mask update the first
thread did resulting in the channel not being marked read on the mask.

Thanks to @Raphfrk for the tip

PS Also stub out ChunkDataMessage as Chunk loading from the server is
coming next.

Signed-off-by: Chris Sanders <zidane@outlook.com>